Move randomMatrix to Utils
